+++ pkg/providers/property/provider.go
@@ -126,6 +126,7 @@ func (p *Subprovider) FrameworkResources() []func() resource.Resource {
 func (p *Subprovider) FrameworkDataSources() []func() datasource.DataSource {
 	return []func() datasource.DataSource{
 		NewCPCodesDataSource,
+		NewHostnameActivationDataSource,
 		NewHostnameActivationsDataSource,
 		NewIncludeDataSource,
 	}